move test classes to lib, move tests to extensions
